Move email things to integrations directory
