+++ tests/test_email/test_email_sender_application.py
@@ -4,9 +4,9 @@ from django.test import override_settings
 from tests.factories import ApplicationFactory, EmailTemplateFactory
 from tests.helpers import patch_method
 from tilavarauspalvelu.enums import EmailType
+from tilavarauspalvelu.integrations.email.email_sender import EmailNotificationSender
 from tilavarauspalvelu.models import Application
 from tilavarauspalvelu.tasks import send_application_handled_email_task, send_application_in_allocation_email_task
-from tilavarauspalvelu.utils.email.email_sender import EmailNotificationSender
 from utils.date_utils import local_datetime
 from utils.sentry import SentryLogger
