+++ tests/test_graphql_api/test_application/test_send.py
@@ -3,7 +3,9 @@ import pytest
 from tests.factories import ApplicationFactory
 from tests.helpers import patch_method
 from tilavarauspalvelu.enums import ApplicationStatusChoice
-from tilavarauspalvelu.utils.email.application_email_notification_sender import ApplicationEmailNotificationSender
+from tilavarauspalvelu.integrations.email.application_email_notification_sender import (
+    ApplicationEmailNotificationSender,
+)
 
 from .helpers import SEND_MUTATION
